requirements.txt
skl2onnx
scikit-onnxruntime
pillow
imutils
https://download.pytorch.org/whl/cu100/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
torchvision==0.3.0